requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
decorator==5.1.1
executing==0.8.3
greenlet==1.1.2
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.29
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
protobuf==4.21.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
